+++ Assets/Enemy Ship.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: bbb5530b9e341054188cd9d9176b178b
+guid: c643aaf643cf70f4f974b1076908aff4
 folderAsset: yes
-timeCreated: 1508877616
+timeCreated: 1511369019
 licenseType: Free
 DefaultImporter:
   userData:
