+++ Assets/Rockets/Materials/New Material.mat.meta
@@ -1,4 +1,7 @@
 fileFormatVersion: 2
-guid: 2f797acb7de6a12469c07ded245ad1f8
+guid: 26c21496097479b42b27af3b590ba56f
 NativeFormatImporter:
+  mainObjectFileID: 0
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
