+++ Assets/Rockets/Materials/New Terrain.asset.meta
@@ -1,4 +1,7 @@
 fileFormatVersion: 2
-guid: dc7e652e741c41543af90d4fab8ad932
+guid: 6eb714332fecb1e4492c8109de56cee7
 NativeFormatImporter:
+  mainObjectFileID: 0
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
