+++ Assets/Rockets/Materials/White.mat.meta
@@ -1,4 +1,7 @@
 fileFormatVersion: 2
-guid: df04dfaf3474ba648bb668f92328f4ce
+guid: 447424b958e61124fad6c110b176a385
 NativeFormatImporter:
+  mainObjectFileID: 0
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
