+++ Assets/Rockets/Scripts/Shooting.cs.meta
@@ -1,8 +1,10 @@
 fileFormatVersion: 2
-guid: 5f57f300688498d4abf0e64bce19b86c
+guid: b7aadbd618cb5e64b8662a8f57aca4e3
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
