+++ Assets/SBSP-A1-nav-combat/Radar/Resources.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: bbb5530b9e341054188cd9d9176b178b
+guid: a0cdf6be1121f684f9d447531e9b91e8
 folderAsset: yes
-timeCreated: 1508877616
+timeCreated: 1511373287
 licenseType: Free
 DefaultImporter:
   userData:
