+++ Assets/SBSP-A1-nav-combat/Shield/Scripts/ShieldController.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 512fa6f0a8c6c934e9956b4999eb853b
-timeCreated: 1510763048
+guid: b4568afffe2a6bc44962bc19dad05ef6
+timeCreated: 1510059716
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
