+++ Assets/SBSP-A1-nav-combat/Shield/Scripts/ShieldModel.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: b05e5e54f50305844bb45e9f277fd785
-timeCreated: 1510763048
+guid: e59d9cc1ab3a411458f6a96ee2549d8d
+timeCreated: 1510059355
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
