+++ Assets/SBSP-A1-nav-combat/Shield/Scripts/ShieldView.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 0ef947e31dd188b41a53435dfafa87c2
-timeCreated: 1510763048
+guid: 551893c2987a92e45aecda681f85ca7f
+timeCreated: 1510059714
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
