--- Assets/SBSP-A1-nav-combat/Shield/Scripts/ShipShield.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5c8c87a84cefb5a458da641feb278a7b
-timeCreated: 1508926548
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
