--- Assets/Shield/Scripts/ShieldController.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: b4568afffe2a6bc44962bc19dad05ef6
-timeCreated: 1510059716
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
