--- Assets/Shield/Scripts/ShieldModel.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: e59d9cc1ab3a411458f6a96ee2549d8d
-timeCreated: 1510059355
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
