--- Assets/Shield/Scripts/ShieldView.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 551893c2987a92e45aecda681f85ca7f
-timeCreated: 1510059714
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
