+++ Assets/_Scenes/MainSceneV002.unity.meta
@@ -1,10 +1,8 @@
 fileFormatVersion: 2
-guid: b5a885063af65774887f132c1399e2bf
-folderAsset: yes
-timeCreated: 1510059375
+guid: d282d8e1df7192c46a58c394e5716420
+timeCreated: 1511366553
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
