+++ Assets/_Scenes/MainSceneV003.unity.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: bbb5530b9e341054188cd9d9176b178b
-folderAsset: yes
-timeCreated: 1508877616
+guid: 5f871979545428b4cad122aed3d5d742
+timeCreated: 1511369019
 licenseType: Free
 DefaultImporter:
   userData:
